[fix] id의 타입이 Long 이므로 findById 의 인자를 Long으로 수정

+++ src/main/java/kr/codesqaud/cafe/domain/user/repository/UserRepository.java
@@ -10,7 +10,7 @@ public interface UserRepository {
 
 	List<User> findAll();
 
-	User findById(Integer id);
+	User findById(Long id);
 
 	void update(User user);
 }
